chore: removal of alerts registry and legacy UIs
BREAKING CHANGE:  Alerts registry was superseded by new overlay alert. Legacy UI was removed in favor of dash
