+++ package.json
@@ -17,11 +17,7 @@
   "license": "MIT",
   "dependencies": {
     "@d-fischer/shared-utils": "^3.6.3",
-    "@sogebot/ui-admin": "^78.0.3",
     "@sogebot/ui-helpers": "^3.1.0",
-    "@sogebot/ui-oauth": "~3.1.1",
-    "@sogebot/ui-overlay": "^36.1.0",
-    "@sogebot/ui-public": "^5.0.0",
     "@twurple/api": "7.0.4",
     "@twurple/auth": "7.0.4",
     "@twurple/chat": "7.0.4",
