+++ src/database/migration/postgres/21.x/1678892044037-removeAlertsRegistry.ts
@@ -1,0 +1,14 @@
+import { MigrationInterface, QueryRunner } from 'typeorm';
+
+export class removeAlertsRegistry1678892044037 implements MigrationInterface {
+  name = 'removeAlertsRegistry1678892044037';
+
+  public async up(queryRunner: QueryRunner): Promise<void> {
+    await queryRunner.query(`DROP TABLE "alert"`);
+  }
+
+  public async down(queryRunner: QueryRunner): Promise<void> {
+    return;
+  }
+
+}
